Added simple AFS2 support. Now it can generate CGSS live music.

+++ DereTore.Application.AcbMaker/Cgss/WaveformTable.cs
@@ -10,7 +10,7 @@ namespace DereTore.Application.AcbMaker.Cgss {
         [UtfField(2)]
         public byte Streaming;
         [UtfField(3)]
-        public ushort NumChannels;
+        public byte NumChannels;
         [UtfField(4)]
         public byte LoopFlag;
         [UtfField(5)]
